Fix userdefaults sequence store issues
